+++ .flake8
@@ -5,6 +5,9 @@ exclude =
     mage_ai/orchestration/db/migrations/versions,
     mage_integrations/mage_integrations/sources/*/tap_*/,
     mage_integrations/mage_integrations/destinations/*/target_*/
+    mage_ai/data_preparation/models/block/dynamic/*
+    mage_ai/data_preparation/models/block/settings/*
+    mage_ai/data/models/outputs/*
 max-line-length = 100
 per-file-ignores =
     mage_ai/orchestration/pipeline_scheduler.py: F401,F403
