+++ mage_ai/data/models/manager.py
@@ -3,6 +3,7 @@ from typing import Any, List, Optional, Union
 import polars as pl
 
 from mage_ai.data.constants import (
+    ChunkKeyTypeUnion,
     InputDataType,
     OutputData,
     RecordBatchGenerator,
@@ -12,9 +13,6 @@ from mage_ai.data.models.base import BaseData
 from mage_ai.data.models.reader import Reader
 from mage_ai.data.models.writer import Writer
 from mage_ai.data.tabular.models import BatchSettings
-from mage_ai.data_preparation.models.block.settings.variables.models import (
-    ChunkKeyTypeUnion,
-)
 from mage_ai.data_preparation.models.utils import infer_variable_type
