+++ mage_ai/data_preparation/models/block/dynamic/variables.py
@@ -9,238 +9,17 @@ from typing import Any, Dict, List, Optional, Tuple, Union
 import pandas as pd
 
 from mage_ai.data.models.outputs.models import BlockOutput
-from mage_ai.data.models.outputs.query import BlockOutputQuery
+from mage_ai.data.models.outputs.query import BlockOutputQuery, DynamicBlockOutputQuery
+from mage_ai.data_preparation.models.block.dynamic.models import (
+    LazyVariableController,
+    LazyVariableSet,
+)
 from mage_ai.data_preparation.models.constants import BlockLanguage, BlockType
 from mage_ai.data_preparation.models.variable import Variable
 from mage_ai.shared.strings import to_ordinal_integers
 
 
-class LazyVariable:
-    def __init__(
-        self,
-        block,
-        variable: Variable,
-        sample: Optional[int] = None,
-        sample_count: Optional[int] = None,
-        skip: bool = False,
-    ):
-        self.block = block
-        self.sample = sample
-        self.sample_count = sample_count
-        self.variable = variable
-
-    @property
-    def is_dynamic(self):
-        from mage_ai.data_preparation.models.block.dynamic.utils import is_dynamic_block
-
-        return is_dynamic_block(self.block)
-
-    def read_data(self):
-        result = self.variable.read_data(
-            sample=self.sample,
-            sample_count=self.sample_count,
-        )
-
-        if self.is_dynamic:
-            return result
-
-        if isinstance(result, list) or isinstance(result, tuple):
-            if len(result) == 1:
-                return result[0]
-
-        return result
-
-    async def read_data_async(self):
-        result = await self.variable.read_data_async(
-            sample=self.sample,
-            sample_count=self.sample_count,
-        )
-
-        if self.is_dynamic:
-            return result
-
-        if isinstance(result, list) or isinstance(result, tuple):
-            if len(result) == 1:
-                return result[0]
-
-        return result
-
-
-class LazyVariableSet(Sequence):
-    def __init__(
-        self,
-        block,
-        variable_objects: List[Variable],
-        logger: Logger = None,
-        logging_tags: Dict = None,
-        **kwargs,
-    ):
-        self.block = block
-        self.lazy_variables = [
-            LazyVariable(
-                block,
-                variable_object,
-                **kwargs,
-            )
-            for variable_object in variable_objects
-        ]
-        self.logger = logger
-        self.logging_tags = logging_tags
-
-    def __getitem__(self, index: int):
-        if index >= len(self.lazy_variables):
-            return {}
-        return self.lazy_variables[index]
-
-    def __iter__(self):
-        for lazy_variable in self.lazy_variables:
-            yield lazy_variable
-
-    def __len__(self):
-        return len(self.lazy_variables)
-
-    @property
-    def is_dynamic(self):
-        from mage_ai.data_preparation.models.block.dynamic.utils import is_dynamic_block
-
-        return is_dynamic_block(self.block)
-
-    @property
-    def lazy_child_data(self) -> Union[List[LazyVariable], LazyVariable]:
-        if len(self) == 2:
-            return self[0]
-        return self.lazy_variables
-
-    @property
-    def lazy_metadata(self) -> Optional[LazyVariable]:
-        if len(self) == 2:
-            return self[1]
-        return None
-
-    def read_child_data(self) -> Any:
-        if not isinstance(self.lazy_child_data, pd.DataFrame) and not self.lazy_child_data:
-            return None
-
-        if isinstance(self.lazy_child_data, list):
-            return [self.read_lazy_variable(data) for data in self.lazy_child_data]
-
-        return (
-            self.read_lazy_variable(self.lazy_child_data)
-            if self.lazy_child_data is not None
-            else None
-        )
-
-    def read_metadata(self) -> Any:
-        return self.read_lazy_variable(self.lazy_metadata) if self.lazy_metadata else {}
-
-    def read_lazy_variable(self, lazy_variable: LazyVariable) -> Any:
-        return lazy_variable.read_data()
-
-    async def read_lazy_variable_async(
-        self,
-        lazy_variable: Union[List[LazyVariable], LazyVariable],
-    ) -> Any:
-        if isinstance(lazy_variable, list):
-            return await asyncio.gather(*[lv.read_data_async() for lv in lazy_variable])
-        elif lazy_variable:
-            return await lazy_variable.read_data_async()
-
-    def read_data(self) -> Tuple[Any, Any]:
-        metadata = self.read_metadata()
-        if metadata is None:
-            metadata = {}
-        return (
-            self.read_child_data(),
-            metadata,
-        )
-
-    async def read_data_async(self) -> Tuple[Optional[Any], Dict]:
-        pair = tuple()
-        if self.lazy_child_data:
-            pair += (await self.read_lazy_variable_async(self.lazy_child_data),)
-        else:
-            pair += (None,)
-
-        if self.lazy_metadata:
-            pair += (await self.read_lazy_variable_async(self.lazy_metadata),)
-        else:
-            pair += ({},)
-
-        return pair
-
-
-class LazyVariableController(Sequence):
-    def __init__(self, block, lazy_variable_sets: List[LazyVariableSet]):
-        self.block = block
-        self.lazy_variable_sets = lazy_variable_sets
-
-    def __getitem__(self, index: int):
-        return self.lazy_variable_sets[index]
-
-    def __iter__(self):
-        for lazy_variable_set in self.lazy_variable_sets:
-            yield lazy_variable_set
-
-    def __len__(self):
-        return len(self.lazy_variable_sets)
-
-    @property
-    def is_dynamic(self):
-        from mage_ai.data_preparation.models.block.dynamic.utils import is_dynamic_block
-
-        return is_dynamic_block(self.block)
-
-    def render(
-        self,
-        child_dynamic_block_index: Optional[int] = None,
-        dynamic_block_index: Optional[int] = None,
-        lazy_load: bool = False,
-    ) -> List[Union[Tuple[Optional[Any], Dict], List[LazyVariableSet]]]:
-        arr = self.lazy_variable_sets
-
-        if child_dynamic_block_index is not None:
-            index = child_dynamic_block_index % len(self)
-            lazy_variable_set = arr[index]
-            child_data, metadata = lazy_variable_set.read_data()
-
-            if self.is_dynamic:
-                if isinstance(child_data, pd.DataFrame):
-                    index = child_dynamic_block_index % len(child_data.index)
-                    child_data = child_data.iloc[index : index + 1]
-                else:
-                    index = child_dynamic_block_index % len(child_data)
-                    child_data = child_data[index]
-                metadata = metadata[index] if len(metadata) > index else {}
-
-            return [child_data, metadata]
-
-        if dynamic_block_index is not None:
-            arr = arr[dynamic_block_index : dynamic_block_index + 1]
-
-        if lazy_load:
-            return arr
-
-        return [lazy_variable_set.read_data() for lazy_variable_set in arr]
-
-    async def render_async(
-        self,
-        dynamic_block_index: Optional[int] = None,
-        lazy_load: bool = False,
-    ) -> List[Union[Tuple[Optional[Any], Dict], List[LazyVariableSet]]]:
-        arr = self.lazy_variable_sets
-
-        if dynamic_block_index is not None:
-            arr = arr[dynamic_block_index : dynamic_block_index + 1]
-
-        if lazy_load:
-            return arr
-
-        return await asyncio.gather(
-            *[lazy_variable_set.read_data_async() for lazy_variable_set in arr],
-        )
-
-
-def get_dynamic_child_block_indexes(
+def __get_dynamic_child_block_indexes(
     block,
     execution_partition: str = None,
 ) -> List[int]:
@@ -264,107 +43,23 @@ def get_dynamic_child_block_indexes(
     return [i for i in range(count)]
 
 
-def get_variable_objects(
-    block,
-    execution_partition: str = None,
-    dynamic_block_index: int = None,
-) -> List[Variable]:
-    """
-    If dynamic_block_index, get the folder names in the nested folder named after the
-    dynamic_block_index (e.g. 0):
-        0/
-            output_0/
-            output_1/
-
-    Or else:
-        output_0/
-        output_1/
-    """
-
-    """
-    If block is a dynamic child block, get the variable objects specifically in the directory
-    named after the dynamic_block_index:
-        0/
-            output_0/
-    """
-    block_uuid = block.uuid
-    if dynamic_block_index is not None:
-        block_uuid = os.path.join(block_uuid, str(dynamic_block_index))
-
-    output_query = BlockOutputQuery(block=block, block_uuid=block_uuid, pipeline=block.pipeline)
-
-    def __sort(block_output: BlockOutput):
-        if block_output is None or block_output.variable is None:
-            return -96, -96
-
-        return (
-            to_ordinal_integers(block_output.variable.block_dir_name)[0],
-            block_output.variable.uuid,
-        )
-
-    return [
-        output.variable
-        for output in output_query.fetch(
-            clean_block_uuid=dynamic_block_index is None,
-            scan_filter=lambda variable_uuid: variable_uuid != '',
-            partition=execution_partition,
-            sort=__sort,
-        )
-    ]
-
-
 def delete_variable_objects_for_dynamic_child(
     block,
-    dynamic_block_index: int = None,
-    execution_partition: str = None,
-) -> None:
-    variable_objects = get_variable_objects(
-        block,
-        dynamic_block_index=dynamic_block_index,
-        execution_partition=execution_partition,
-    )
-    if variable_objects:
-        for variable_object in variable_objects:
-            variable_object.delete()
-
-
-def __get_all_variable_objects_for_dynamic_child(
-    block,
+    dynamic_block_index: Optional[int] = None,
     execution_partition: Optional[str] = None,
-) -> List[List[Variable]]:
+) -> None:
     """
-    This method will get the nested outputs (output_0) in every numeric folder
-    named after the dynamic_block_index (e.g. 0/).
+    Used by:
+        block.store_variables
     """
-    variable_objects_arr = []
-
-    indexes = get_dynamic_child_block_indexes(block, execution_partition=execution_partition)
-    for dynamic_block_index in indexes:
-        # 0/output_0,
-        # 0/output_1,
-        # 0/output_2,
-
-        # get_variable_objects returns a list of outputs (e.g. [output_0, output_1, output_2])
-        arr = get_variable_objects(
-            block,
-            execution_partition=execution_partition,
-            dynamic_block_index=dynamic_block_index,
-        )
-
-        def __sort(variable_object: str):
-            if not variable_object:
-                return -96, -96
-            return (
-                to_ordinal_integers(variable_object.block_dir_name)[0],
-                variable_object.uuid,
-            )
-
-        variable_objects_arr.append(sorted(arr, key=__sort))
-
-    return variable_objects_arr
+    output_query = DynamicBlockOutputQuery(block=block)
+    block_outputs = output_query.fetch(dynamic_block_index, partition=execution_partition)
+    if block_outputs:
+        for output in block_outputs:
+            output.variable.delete()
 
 
-async def get_outputs_async(
+async def __get_outputs_async(
     block,
     execution_partition: Optional[str] = None,
     dynamic_block_index: Optional[int] = None,
@@ -386,7 +81,7 @@ async def get_outputs_async(
     ])
 
 
-def get_outputs(
+def __get_outputs(
     block,
     execution_partition: str = None,
     dynamic_block_index: int = None,
@@ -411,7 +106,12 @@ def get_outputs(
 def get_outputs_for_dynamic_block(
     block, **kwargs
 ) -> List[Optional[Union[Dict, int, str, pd.DataFrame, Any]]]:
-    values = get_outputs(block, **kwargs)
+    """
+    mage_ai/data_preparation/models/block/dynamic/child.py
+    mage_ai/data_preparation/models/block/dynamic/utils.py
+    block.__get_outputs_async (old)
+    """
+    values = __get_outputs(block, **kwargs)
 
     if BlockLanguage.SQL == block.language:
         return [values[0] if len(values) == 1 else values, None]
@@ -427,7 +127,10 @@ def get_outputs_for_dynamic_block(
 async def get_outputs_for_dynamic_block_async(
     block, **kwargs
 ) -> List[Optional[Union[Dict, int, str, pd.DataFrame, Any]]]:
-    values = await get_outputs_async(block, **kwargs)
+    """
+    block.__get_outputs_async (old)
+    """
+    values = await __get_outputs_async(block, **kwargs)
 
     if BlockLanguage.SQL == block.language:
         return values[0] if len(values) == 1 else values, None
@@ -448,6 +151,11 @@ def get_outputs_for_dynamic_child(
     sample: bool = False,
     sample_count: Optional[int] = None,
 ) -> List[Tuple[List[Union[Dict, int, str, pd.DataFrame]], List[Dict]]]:
+    """
+    mage_ai/data_preparation/models/block/dynamic/child.py
+    block.get_outputs (old)
+    block.__get_outputs_async (old)
+    """
     # List[List[Variable]]
     list_of_lists_of_variables = __get_all_variable_objects_for_dynamic_child(
         block,
@@ -479,6 +187,10 @@ def fetch_input_variables_for_dynamic_upstream_blocks(
     global_vars: Dict = None,
     **kwargs,
 ) -> Tuple[List, List, List]:
+    """
+    block.fetch_input_variables
+    block.interpolate_content
+    """
     from mage_ai.data_preparation.models.block.dynamic.utils import (
         is_dynamic_block,
         is_dynamic_block_child,
